Refact orc casting tests.
