+++ integration_tests/src/main/python/orc_test.py
@@ -645,26 +645,6 @@ def test_orc_scan_with_aggregate_no_pushdown_on_col_partition(spark_tmp_path, ag
                 conf=_orc_aggregate_pushdown_enabled_conf)
 
 
-@pytest.mark.parametrize('offset', [1,2,3,4], ids=idfn)
-@pytest.mark.parametrize('reader_confs', reader_opt_confs, ids=idfn)
-@pytest.mark.parametrize('v1_enabled_list', ["", "orc"])
-def test_read_type_casting_integral(spark_tmp_path, offset, reader_confs, v1_enabled_list):
-    int_gens = [boolean_gen] + integral_gens
-    gen_list = [('c' + str(i), gen) for i, gen in enumerate(int_gens)]
-    data_path = spark_tmp_path + '/ORC_DATA'
-    with_cpu_session(
-        lambda spark: gen_df(spark, gen_list).write.orc(data_path))
-
-    # build the read schema by a left shift of int_gens
-    shifted_int_gens = int_gens[offset:] + int_gens[:offset]
-    rs_gen_list = [('c' + str(i), gen) for i, gen in enumerate(shifted_int_gens)]
-    rs = StructGen(rs_gen_list, nullable=False).data_type
-    all_confs = copy_and_update(reader_confs,
-        {'spark.sql.sources.useV1SourceList': v1_enabled_list})
-    assert_gpu_and_cpu_are_equal_collect(
-        lambda spark: spark.read.schema(rs).orc(data_path),
-        conf=all_confs)
-
 def test_orc_read_count(spark_tmp_path):
     data_path = spark_tmp_path + '/ORC_DATA'
     orc_gens = [int_gen, string_gen, double_gen]
